requirements.txt
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
seaborn==0.13.2
matplotlib==3.8.2
python-dotenv==1.0.0
jinja2==3.1.2
kaleido==0.2.1